improvement: add secrets in db confiugration (#4522)
